chore(*): Move whole logic for input type change in base class instead of in a service so that app component does not depend on it.
