+++ src/app/app.component.ts
@@ -1,6 +1,5 @@
 import { DOCUMENT } from "@angular/common";
 import { Component, HostListener, Inject, OnInit } from "@angular/core";
-import { GlobalInputTypeService } from "./services/global-input-type.service";
 
 @Component({
     selector: "app-root",
@@ -15,7 +14,7 @@ export class AppComponent implements OnInit {
     private typefacesLoaded = ["Titillium Web", "Roboto"];
     private typefaceUrl = "https://fonts.googleapis.com/css?family=";
 
-    constructor(@Inject(DOCUMENT) private document: Document, private typeservice: GlobalInputTypeService) {}
+    constructor(@Inject(DOCUMENT) private document: Document) {}
 
     public ngOnInit() {
         this.createThemeStyle();
@@ -25,11 +24,6 @@ export class AppComponent implements OnInit {
     private onMessage(e: MessageEvent) {
         if (e.origin === e.data.origin && typeof e.data.themeStyle === "string") {
             this.styleElem.textContent = e.data.themeStyle;
-            const themeName: string = e.data.themeName;
-            if (themeName) {
-                this.applyInputTypeForTheme(themeName);
-            }
-
             const typeface = window.getComputedStyle(this.document.body).fontFamily.replace(/\"/g, "");
             if (!(typeface.match(/,/g) || []).length &&
                 !this.typefacesLoaded.includes(typeface)) {
@@ -40,16 +34,6 @@ export class AppComponent implements OnInit {
             this.document.body.classList.remove(this.theme);
             this.document.body.classList.add(e.data.theme);
             this.theme = e.data.theme;
-        }
-    }
-
-    private applyInputTypeForTheme(theme: string) {
-        const themeName = theme.split("-")[1];
-        switch (themeName) {
-            case "theme": this.typeservice.setType("line"); break;
-            case "fluent": this.typeservice.setType("fluent"); break;
-            case "bootstrap": this.typeservice.setType("bootstrap"); break;
-            default: break;
         }
     }
