+++ src/app/data-entries/input-group/base-input.component.ts
@@ -1,26 +1,29 @@
-import { AfterViewInit, Directive, QueryList, ViewChildren } from "@angular/core";
+import { AfterViewInit, Directive, HostListener, QueryList, ViewChildren } from "@angular/core";
 import { IgxInputGroupComponent } from "igniteui-angular";
-import { GlobalInputTypeService } from "../../services/global-input-type.service";
 
 @Directive({
     selector: "input-group-sample-base"
 })
-export class BaseInputGroupSampleComponent implements AfterViewInit {
+export class BaseInputGroupSampleComponent {
 
     public type = null;
     @ViewChildren(IgxInputGroupComponent) public inputGroups: QueryList<IgxInputGroupComponent>;
 
-    constructor(private inputType: GlobalInputTypeService) {
-        inputType.typeName.subscribe((newValue) => {
-            if (newValue) {
-                this.type = newValue;
-                this.setInputType(this.type);
-            }
-        });
+    @HostListener("window:message", ["$event"])
+    private onMessage(e: MessageEvent) {
+        const themeName: string = e.data.themeName;
+        if (themeName) {
+            this.applyInputTypeForTheme(themeName);
+        }
     }
-    public ngAfterViewInit() {
-        if (this.type) {
-            this.setInputType(this.type);
+
+    private applyInputTypeForTheme(theme: string) {
+        const themeName = theme.split("-")[1];
+        switch (themeName) {
+            case "theme": this.setInputType("line"); break;
+            case "fluent": this.setInputType("fluent"); break;
+            case "bootstrap": this.setInputType("bootstrap"); break;
+            default: break;
         }
     }
